try-fix: 简单的重入检查、尝试修复小尺寸icmp报文引起的ping-rtt不回显

+++ init.sh
@@ -3,13 +3,14 @@
 # SPDX-LICENSE-IDENTIFIER: GPL2.0
 # (C) All rights reserved. Author: <kisfg@hotmail.com> in 2025
 # Created at 2025年07月06日 星期日 18时04分20秒
-# Last modified at 2025年07月30日 星期三 01时36分00秒
+# Last modified at 2025年07月30日 星期三 13时31分51秒
 set -e
 
 # github
 raw_github='https://raw.githubusercontent.com'
 main_github='https://github.com'
 release_path='releases/download'
+url_prefix="$main_github"
 
 # 字体包名称
 mononame="JetBrainsMono-2.304"
@@ -18,20 +19,29 @@ lxgwname="lxgw-wenkai-v1.520"
 # 这两个只有 zip
 mono_zip="$mononame.zip"
 fira_zip="$firaname.zip"
-lxgw_tar="$lxgwname.tar.gz"
+lxgw_zip="$lxgwname.zip"
 # 有点难办
-sha256_list=('5ecb50e9f5aa644d0aebba93881183f0a7b9aaf829bac9dbadaf348f557e0029' 'b9caa260fde3cb5681711f91dbfc2d6ec7ecf2fabbf92cef4432fc19c9a73816' '25d806b8ac55e21cddd3a1fdcbc929d3a232a1cac277ae606158824d803d2d09')
+sha256_list=(
+	'5ecb50e9f5aa644d0aebba93881183f0a7b9aaf829bac9dbadaf348f557e0029'
+	'b9caa260fde3cb5681711f91dbfc2d6ec7ecf2fabbf92cef4432fc19c9a73816'
+	'25d806b8ac55e21cddd3a1fdcbc929d3a232a1cac277ae606158824d803d2d09'
+)
 
 # 本地配置
 vimdir="$HOME/.vim"
 color_path="$vimdir/colors"
 init_dir="$vimdir/autoload"
 plugman="$init_dir/plug.vim"
-fonts_dir="$vimdir/fonts/"
+fonts_dir="$HOME/.fonts/"
 
-url_prefix="$main_github"
+
 # TODO: 如何获取镜像主机名单？
-famous_mirrors=('wget.la' 'gh-proxy.com' 'gh.xmly.dev' 'gh.llkk.cc')
+famous_mirrors=(
+	'wget.la'
+	'gh-proxy.com'
+	'gh.xmly.dev'
+	'gh.llkk.cc'
+)
 main_mirror=''
 
 # ping 所有给定的镜像站，取时长最小的一个
@@ -41,12 +51,16 @@ function _probe() {
 	rtt_dict=()
 	for cur_mirror in ${famous_mirrors[@]}; do
 		mid_rtt_val=`ping -c 2 $cur_mirror`
-		echo $mid_rtt_val
 		rtt_val=`echo $mid_rtt_val | grep '^rtt' | awk -F'/' '{ print $6 }'`
-		rtt+=($rtt_val)
-		if [ -n "$rtt_val" ]; then
-			rtt_dict[$rtt_val]="$cur_mirror"
+		# issue: ping包比较小就不会携带rtt信息
+		# 2 packets transmited, 0 received, ... , time 1022ms
+		if [[ "$rtt_Val" == "" ]]; then
+			rtt_val=`echo $mid_rtt_val | grep 'time [0-9]\+ms$' | awk -F' ' ' {print $2 } '`
 		fi
+		# 否则认为站点不可达
+		[[ "$rtt_Val" == "" ]] && rtt_val=31415926535897932384626433
+		rtt+=("$rtt_val")
+		[ -n "$rtt_val" ] && rtt_dict[$rtt_val]="$cur_mirror"
 	done
 	rtt=`printf "%s\n" ${rtt[@]} | sort -n | head -n 1 | awk -F'\n' '{ print $1 }'`
 	main_mirror=${rtt_dict[$rtt]}
@@ -75,10 +89,9 @@ function alter_src_via_mirror() {
 # 入参:  压缩包名称 url
 function _detect_font() {
 	font_urls=($@)
-	# TODO: 其实可以全部用 unzip 的
 	fontname_list=("$mononame" "$firaname" "$lxgwname")
-	op_list=('unzip' 'unzip' 'tar -xf')
-	tar_list=("$mono_zip" "$fira_zip" "$lxgw_tar")
+	# op_list=('unzip' 'unzip' 'unzip')
+	tar_list=("$mono_zip" "$fira_zip" "$lxgw_zip")
 	for ((i=0; i<${#font_urls[@]}; i++)); do
 		if [[ ${fontname_list[i]} != '' &&  -d "./${fontname_list[i]}" ]]; then
 			ret=`tar -c ${fontname_list[i]} | sha256sum | awk -F' ' ' { print $1 } '`
@@ -86,11 +99,12 @@ function _detect_font() {
 			[[ "$ret" == ${sha256_list[i]} ]] && continue
 		elif [[ -f ${tar_list[i]} ]]; then
 			# 不存在但有tar/zip
-			${op_list[i]} ${tar_list[i]} && rm ${tar_list[i]}
+			# ${op_list[i]} ${tar_list[i]} && rm ${tar_list[i]}
+			unzip ${tar_list[i]} && rm ${tar_list[i]}
 			continue
 		fi
-		wget ${font_urls[i]}
-		${op_list[i]} ${tar_list[i]} && rm ${tar_list[i]}
+		wget ${font_urls[i]} && unzip ${tar_list[i]} && rm ${tar_list[i]}
+		# ${op_list[i]} ${tar_list[i]} && rm ${tar_list[i]}
 	done
 }
 
@@ -98,17 +112,13 @@ function get_fonts() {
 	# TODO: 如果能在这里换最新的字体也不错
 	jetbrain="$url_prefix/JetBrains/JetBrainsMono/$release_path/v2.304/$mono_zip"
 	firacode="$url_prefix/tonsky/FiraCode/$release_path/6.2/$fira_zip"
-	lxgw="$url_prefix/lxgw/LxgwWenkai/$release_path/v1.520/$lxgw_tar"
+	lxgw="$url_prefix/lxgw/LxgwWenkai/$release_path/v1.520/$lxgw_zip"
 	curr_dir=`pwd`
+
+	# 到HOME目录
 	mkdir -p "$fonts_dir" && cd "$fonts_dir"
 	link_list=("$jetbrain" "$firacode" "$lxgw")
 	"_detect_font" ${link_list[@]}
-
-	# 拷贝一份到HOME目录
-	# TODO: 早知如此何必当初？
-	mkdir -p $HOME/.fonts/
-	cp -r $fonts_dir* $HOME/.fonts/
-
 	fc-cache -f -v
 	ret=`fc-list | grep -Ei "$lxgwname|$firaname|$mononame"`
 	if [[ "$ret" == '' || "$?" != 0 ]]; then
@@ -120,15 +130,16 @@ function get_fonts() {
 }
 
 function get_color_scheme() {
+	[[ -d "$color_path" && -f "$color_path/gruvbox.vim" ]] && return # 加入判断，避免重复下载
 	mkdir -p "$color_path"
-	# TODO: 加入判断，避免重复下载
 	wget "$raw_github/morhetz/gruvbox/master/colors/gruvbox.vim" -O "$color_path/gruvbox.vim"
 }
 
 function get_plug_manager() {
-	# TODO: 加入判断，避免重复下载
+	[ -f "$plugman" ] && return # 加入判断，避免重复下载
 	curl -fLo "$plugman" --create-dirs \
 		 "$raw_github/junegunn/vim-plug/baa66bcf349a6f6c125b0b2b63c112662b0669e1/plug.vim"
+	# 不可将以下关系合并到上面的判断
 	[[ "$res" == '' ]] && return
 	# 需要备份plugman，防止意外
 	# 后面自己删
